changed index length

+++ database/migrations/2020_10_14_211856_created_foreign_option_column.php
@@ -15,7 +15,7 @@ class CreatedForeignOptionColumn extends Migration
     {
         Schema::table('price_option', function(Blueprint $table)
         {
-            $table->bigIncrements('foreign_id')->unsigned()->nullable();
+            $table->integer('foreign_id')->unsigned()->nullable();
             $table->foreign('foreign_id')->references('id')->on('foreign_option');
         });
     }
